+++ packages/uba/CHANGELOG.md
@@ -15,6 +15,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.20.16] - 2021-06-14
+### Added
+ - `uba_usercertificate` - added attributes `certParsed: Json` and `isForSigning: boolean`.
+   Both are sets on certificate binary changes.
+   Consider to use `uba_user.addCertificate` added by `@ub-d/crypto-api` method to add a binary certificate
+   
+ - `isForSigning` attribute added to the field list of Users->Certificates shortcut;
+
+ - `uba_usercertificate.getCertificate` method. Can return:
+    - any certificate if called as `rest/uba_usercertificate/getCertificate?ID=1231`
+    - SIGNING certificate for current user if called w/o ID parameter `rest/uba_usercertificate/getCertificate`
+    - if called as UBQL, certificate returned as base64 encoded string
+
+### Changed
+ - `uba_usercertificate` form is rewritten to Vue
+
+### Removed
+ - `@unitybase/uba` does not depend on `asn1js` and `pkijs` packages anymore, certificate parser is moved to server side `@ub-d/crypto-api` model  
+
 ## [5.20.15] - 2021-05-24
 ## [5.20.14] - 2021-05-13
 ### Fixed